requirements.txt
streamlit
google-genai
python-dotenv
opencv-python
numpy
pillow
diffusers
transformers
accelerate
torch
pycountry
requests